[WIP] add hit

+++ Server/Utils/Timing.cs
@@ -14,12 +14,12 @@ namespace Server.Utils
 		Stopwatch _stopWatch;
 		Timer _timer;
 		Timing() { }
-		public static long DeltaTime { get; private set; }
+		public static float DeltaTime { get; private set; }
 		public static ulong Fps { get; private set; }
 		public static void Init()
 		{
 			Fps = 0;
-			DeltaTime = 1 / 60;
+			DeltaTime = 1 / 60f;
 		}
 		public static void OnNewFrameStart(long deltaTime)
 		{
